Move test helpers to dedicated package
